[Feature] #10 - 코드 네이밍 수정
